python-brain/pyproject.toml
[project]
name = "python-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.43.0",
]
